fix : 직무, 유형 예외 처리 수정

+++ inlab-api/src/main/java/kr/inlab/www/service/PositionServiceImpl.java
@@ -1,6 +1,5 @@
 package kr.inlab.www.service;
 
-import javax.management.relation.RoleNotFoundException;
 import kr.inlab.www.common.exception.*;
 import kr.inlab.www.common.type.RoleType;
 import kr.inlab.www.common.util.PagingUtil;
@@ -14,11 +13,7 @@ import kr.inlab.www.entity.Position;
 import kr.inlab.www.entity.PositionLevel;
 import kr.inlab.www.entity.Role;
 import kr.inlab.www.entity.User;
-import kr.inlab.www.repository.PositionLevelRepository;
-import kr.inlab.www.repository.PositionRepository;
-import kr.inlab.www.repository.QuestionRepository;
-import kr.inlab.www.repository.RoleRepository;
-import kr.inlab.www.repository.UserRepository;
+import kr.inlab.www.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -27,6 +22,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.management.relation.RoleNotFoundException;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Map;
@@ -40,8 +36,8 @@ public class PositionServiceImpl implements PositionService {
     private final QuestionRepository questionRepository;
     private final UserRepository userRepository;
     private final PositionLevelRepository positionLevelRepository;
+    private final QuestionTypeRepository questionTypeRepository;
     private final PositionLevelService positionLevelService;
-    private final QuestionLevelService questionLevelService;
     private final RoleRepository roleRepository;
 
     @Override
@@ -57,8 +53,6 @@ public class PositionServiceImpl implements PositionService {
         Role role = roleRepository.findByRoleType(RoleType.ROLE_ADMIN).orElseThrow(RoleNotFoundException::new);
         List<User> adminUserList = userRepository.findAllByRoles(role);
         positionLevelService.createAdminPositionLevel(adminUserList, position);
-
-
     }
 
     @Override
@@ -75,10 +69,13 @@ public class PositionServiceImpl implements PositionService {
     public void deletePosition(Integer positionId) {
         Position position = positionRepository.findById(positionId)
                 .orElseThrow(PositionNotFoundException::new);
+
+        if (questionRepository.existsByPosition_PositionId(positionId) ||
+                questionTypeRepository.existsByPosition_PositionId(positionId))
+            throw new PositionDeleteNotAllowedException();
+
         positionLevelRepository.deleteAllByPosition(position);
         positionRepository.delete(position);
-
-
     }
 
     @Override
@@ -87,8 +84,11 @@ public class PositionServiceImpl implements PositionService {
         Position position = positionRepository.findById(positionId)
                 .orElseThrow(PositionNotFoundException::new);
 
-        if (questionRepository.countByPosition(position) > 0)
+        if (questionRepository.existsByPosition_PositionId(positionId))
             throw new PositionUpdateNotAllowedException();
+
+        if (positionRepository.existsByPositionName(requestDto.getPositionName()))
+            throw new PositionAlreadyExistsException();
 
         position.updateName(requestDto.getPositionName());
     }
